add PR changes and left comments on suggested changes
