+++ src/styles/styles/index.js
@@ -1,9 +1,3 @@
-import {Platform} from 'react-native';
-
-const styles = Platform.select({
-    ios: require('./styles.ios').default,
-    default: require('./styles.native').default,
-    android: require('./styles.android').default,
-});
+import styles from './baseStyles';
 
 export default styles;
