+++ src/styles/variables/index.js
@@ -1,8 +1,3 @@
-import {Platform} from 'react-native';
-
-const variables = Platform.select({
-    web: require('./variables.web').default,
-    default: require('./variables.native').default,
-});
+import variables from './baseVariables';
 
 export default variables;
